Lab - Dolphin and Goon Rigs
learned how to use the rig/animation in relation to the bones and premade animation timelines. This gives me such big LittleBigPlanet vibes, in that game you can emote and have your Sackboy move their arms with the controller joycons, and also they had to walk on all kinds of slopes and shapes, jumping and surviving explosions in cartoonish physics. I can see how procedural animation is used with the concepts we learned with this, and I guess I just wanted to say that I think it's really cool to learn this stuff and seems really fun